requirements.txt
numpy==1.26.0
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
statsmodels==0.14.4
pmdarima==2.0.4
prophet==1.1.6
scikit-learn==1.5.2
python-dotenv==1.0.1
google-cloud-bigquery==3.25.0
google-auth==2.35.0